integrate APIservice
