+++ src/app/tmdbApi.js
@@ -11,6 +11,26 @@ const tmdbApi = {
       throw error;
     }
   },
+  getMovieDetails: async (movieId) => {
+    const url = `movie/${movieId}`;
+    try {
+      const data = await apiService.get(url);
+      return data;
+    } catch (error) {
+      console.log(error);
+      throw error;
+    }
+  },
+  getMovieGenres: async () => {
+    const url = `genre/movie/list`;
+    try {
+      const data = await apiService.get(url);
+      return data;
+    } catch (error) {
+      console.log(error);
+      throw error;
+    }
+  },
 };
 
 export default tmdbApi;
